// packages/api/src/auth/enums/auth-role.enum.ts
export enum AuthRole {
  USER = 'USER',
  QUANTIFIER = 'QUANTIFIER',
  FORWARDER = 'FORWARDER',
  ADMIN = 'ADMIN',
  ROOT = 'ROOT',
  API_KEY_READWRITE = 'API_KEY_READWRITE',
  API_KEY_READ = 'API_KEY_READ',
  API_KEY_DISCORD_BOT = 'API_KEY_DISCORD_BOT',
  API_KEY_SETUP_WEB = 'API_KEY_SETUP_WEB',
}